Removed redundant old App component. Main renamed to App
